Update exercise and add slides
